// frontend/src/gen/api/src/models/MediaDtoGenericList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NewLevel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MediaDto } from './MediaDto';
import {
    MediaDtoFromJSON,
    MediaDtoFromJSONTyped,
    MediaDtoToJSON,
    MediaDtoToJSONTyped,
} from './MediaDto';

/**
 * 
 * @export
 * @interface MediaDtoGenericList
 */
export interface MediaDtoGenericList {
    /**
     * 
     * @type {number}
     * @memberof MediaDtoGenericList
     */
    totalCount?: number;
    /**
     * 
     * @type {Array<MediaDto>}
     * @memberof MediaDtoGenericList
     */
    items?: Array<MediaDto> | null;
}

/**
 * Check if a given object implements the MediaDtoGenericList interface.
 */
export function instanceOfMediaDtoGenericList(value: object): value is MediaDtoGenericList {
    return true;
}

export function MediaDtoGenericListFromJSON(json: any): MediaDtoGenericList {
    return MediaDtoGenericListFromJSONTyped(json, false);
}

export function MediaDtoGenericListFromJSONTyped(json: any, ignoreDiscriminator: boolean): MediaDtoGenericList {
    if (json == null) {
        return json;
    }
    return {
        
        'totalCount': json['totalCount'] == null ? undefined : json['totalCount'],
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(MediaDtoFromJSON)),
    };
}

export function MediaDtoGenericListToJSON(json: any): MediaDtoGenericList {
    return MediaDtoGenericListToJSONTyped(json, false);
}

export function MediaDtoGenericListToJSONTyped(value?: MediaDtoGenericList | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'totalCount': value['totalCount'],
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(MediaDtoToJSON)),
    };
}